articles_form_comment_page_edit_articles
so this will be updated. not finished. This will have articles form so users can create an articles, this also has it in the nav bar so you can click create articles from nav, a page for the articles once you click on one to view articles. it also has in the articles page a button so they can comment on articles. it also has edit articles feature.