Created more components
Created a Profile Page, the idea is that from the mentor page you can use the Profile Page as a Detail View of a "mentor". If the user is logged in the they will also get access to "update" their profile information.

Created a "ProfileForm" component that allows the user to update their profile information.

I've created some "Interest" icons which are apart of the "InterestIcons" component. The idea with this is that the will go a more opaque colour if those interests have been selected by the mentor.

I went through and cleared up some imports that were not being used so that they don't show errors in Powershell.

I seem to be having some issues with Mapping the MentorCard components. I will be looking into fixing this next.

We may still be missing some functionality from our endpoints. For example, I noticed that "PUT" was not allowed when updating the user information. I am hoping that I just have the incorrect endpoint but I will also bring it up with the group.